New/Updated Embedded End Code
New code for the Embedded end of things. Some progress made mainly transmitting data via usb serial port. As well as some other tweaks. The code to transmit data is looping and continuously transmitting as would be expected from the scope. Not sure as to if the data is accurate or if the format is acceptable???